+++ src/ParallelDomainCollector.cpp
@@ -588,7 +588,7 @@ throw (DCException)
 
 void ParallelDomainCollector::appendDomain(int32_t id,
         const CollectionType& type,
-        uint32_t count,
+        size_t count,
         const char *name,
         const Dimensions domainOffset,
         const Dimensions domainSize,
@@ -600,9 +600,9 @@ throw (DCException)
 
 void ParallelDomainCollector::appendDomain(int32_t id,
         const CollectionType& type,
-        uint32_t count,
-        uint32_t offset,
-        uint32_t striding,
+        size_t count,
+        size_t offset,
+        size_t striding,
         const char *name,
         const Dimensions domainOffset,
         const Dimensions domainSize,
