+++ src/include/IParallelDataCollector.hpp
@@ -220,15 +220,15 @@ namespace DCollector
         
         virtual void append(int32_t id,
                 const CollectionType& type,
-                uint32_t count,
+                size_t count,
                 const char *name,
                 const void *data) = 0;
 
         virtual void append(int32_t id,
                 const CollectionType& type,
-                uint32_t count,
-                uint32_t offset,
-                uint32_t stride,
+                size_t count,
+                size_t offset,
+                size_t stride,
                 const char *name,
                 const void *data) = 0;
