+++ src/include/ParallelDataCollector.hpp
@@ -425,7 +425,7 @@ namespace DCollector
         
         void append(int32_t id,
                 const CollectionType& type,
-                uint32_t count,
+                size_t count,
                 const char *name,
                 const void *data)
         {
@@ -434,9 +434,9 @@ namespace DCollector
 
         void append(int32_t id,
                 const CollectionType& type,
-                uint32_t count,
-                uint32_t offset,
-                uint32_t stride,
+                size_t count,
+                size_t offset,
+                size_t stride,
                 const char *name,
                 const void *data)
         {
